# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Proyecto Final",
        "description": "Documentación de la API para el proyecto final",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/review": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Almacena una nueva review.",
                "description": "Este método crea una nueva review utilizando los datos proporcionados en la solicitud. Si la review se crea con éxito, devuelve una respuesta JSON con un mensaje de éxito y los detalles de la review creada. Si ocurre algún error durante el proceso de creación de la review, devuelve una respuesta JSON con un mensaje de error.",
                "operationId": "storeReview",
                "requestBody": {
                    "description": "La solicitud HTTP que contiene los datos de la review a crear.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "commerce_username",
                                    "comment",
                                    "note"
                                ],
                                "properties": {
                                    "commerce_username": {
                                        "type": "string",
                                        "example": "nombre_usuario_comercio"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Contenido_del_comentario"
                                    },
                                    "note": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Review creada"
                                        },
                                        "data": {
                                            "properties": {
                                                "commerce_id": {
                                                    "type": "string",
                                                    "example": "ID_del_comercio"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Contenido_del_comentario"
                                                },
                                                "note": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuario inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario inexistente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El usuario $request->commerce_username no es un comercio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El usuario $request->commerce_username no es un comercio"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear la review: mensaje_de_error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear la review: mensaje_de_error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}